+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Enhancements
   - Updated UI styling with StationUI
   - Unsaved changes will now be saved when publishing, instead of discarded
+  - Adds Unpublish button to Page editor
+  - Page editor now shows draft/published status
+
+### Fixes
+
+  - Fixed a bug where MediaLibrary could check for file contents on the wrong node in multi-node deployments
 
 ## 0.3.1 (2024-12-12)